From: "John Cavanaugh" <john-cavanaugh@attbi.com> Subject: Patch for Image::Info (Canon Makernote support) To: <gisle@activestate.com> Date: Wed, 16 Apr 2003 12:04:42 -0700
Gisle,

Attached is a patch that provides support for the Canon makernote in
Image::Info (To support the G1/2/3, S series, D30/D60/10D).  It also
adds in one Fuji camera to the recognized camera list (It happened to be
my old fuji camera before upgrading to a 10D).

Hopefully you can get this rolled into an upcoming release of
Image::Info.

Ive got a buddy with a Kodak camera who is working on its makernote, but
since the docs are unavailable, that one may take some time.  Ill
forward you that one when/if it becomes available.

--John Cavanaugh

[git-p4: depot-paths = "//depot/users/gisle/p/img-info/": change = 78439]